Set RES_EMBED_USE_SHARED for add_subdirectory usage
This variable is used for linkage type in ResEmbed function and it is
defined in res_embedConfig.cmake which is utilized when res embed is
installed as a package in the system. In order to use resembed by add_subdirectory
we need to set this variable ourselves and for it to take effect we need
to set scope to parent scope.